Move SSPI interop
