--- src/Microsoft.Data.SqlClient/netcore/src/Common/Interop/Windows/sspicli/SecPkgContext_Bindings.cs
@@ -1,20 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-#if !NET8_0_OR_GREATER
-
-using System.Runtime.InteropServices;
-
-namespace System.Net
-{
-    // sspi.h
-    [StructLayout(LayoutKind.Sequential)]
-    internal struct SecPkgContext_Bindings
-    {
-        internal int BindingsLength;
-        internal IntPtr Bindings;
-    }
-}
-
-#endif
