+++ src/Microsoft.Data.SqlClient/src/Interop/Windows/SspiCli/SSPIInterface.netcore.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-#if !NET8_0_OR_GREATER
+#if !NETFRAMEWORK && !NET8_0_OR_GREATER
 
 using System.Net.Security;
 using System.Runtime.InteropServices;
